Move serviceOne files into their respective folder
